create  and move addon from , clean up SB6
