+++ .changeset/shiny-sloths-drum.md
@@ -1,7 +1,8 @@
 ---
-'@web/storybook-framework-web-components': minor
+'@web/storybook-addon-mocks': major
 '@web/storybook-builder': minor
-'@web/storybook-utils': minor
+'@web/storybook-framework-web-components': minor
+'@web/storybook-utils': major
 ---
 
 Migrate projects to Storybook 9
