+++ packages/mocks/README.md
@@ -1,8 +1,6 @@
 # @web/mocks
 
-## Deprecation notice: `@web/mocks` is compatible only up to Storybook 8. For Storybook 9, please use `@web/storybook-addons-mocks` instead.
-
-[`MSW`](https://mswjs.io/) integration layer for usage with [`@web/dev-server`](https://modern-web.dev/docs/dev-server/overview/), [`@web/test-runner`](https://modern-web.dev/docs/test-runner/overview/) and [`@web/dev-server-storybook`](https://modern-web.dev/docs/dev-server/plugins/storybook/#mainjs-and-previewjs).
+[`MSW`](https://mswjs.io/) integration layer for usage with [`@web/dev-server`](https://modern-web.dev/docs/dev-server/overview/), [`@web/test-runner`](https://modern-web.dev/docs/test-runner/overview/).
 
 ## Defining mocks
 
@@ -91,175 +89,6 @@ http.get('/api/foo', ({ request, cookies, params }) => {
 - `request` the native `Request` object
 - `cookies` an object based on the request cookies
 - `params` an object based on the request params
-
-## `@web/dev-server`/`@web/dev-server-storybook`/`@web/storybook-builder`
-
-`feature-a/web-dev-server.config.mjs`:
-
-```diff
-+// for Storybook 7+ (@web/storybook-builder)
-+// no need to do anything here
-+// this file is not even needed for "@web/storybook-builder"
-
-import { storybookPlugin } from '@web/dev-server-storybook';
-
-+// for Storybook 6 (@web/dev-server-storybook)
-+import { mockPlugin } from '@web/mocks/plugins.js';
-
-export default {
-  nodeResolve: true,
-  plugins: [
-+    // for Storybook 6 (@web/dev-server-storybook)
-+    mockPlugin(),
-    storybookPlugin({ type: 'web-components' })
-  ],
-};
-```
-
-You can also add the `mockRollupPlugin` to your `.storybook/main.cjs` config for when you're bundling your Storybook to deploy somewhere; your mocks will be deployed along with your Storybook, and will work in whatever environment you deploy them to.
-
-`feature-a/.storybook/main.cjs`:
-
-```diff
-module.exports = {
-  stories: ['../stories/**/*.stories.{js,md,mdx}'],
-+  // for Storybook 7+ (@web/storybook-builder)
-+  // no need to do anything here
-
-+  // for Storybook 6 (@web/dev-server-storybook)
-+  rollupConfig: async config => {
-+    const { mockRollupPlugin } = await import('@web/mocks/plugins.js');
-+    config.plugins.push(mockRollupPlugin());
-+    return config;
-+  },
-};
-```
-
-<details>
-<summary>
-  <b><code>mockRollupPlugin</code> configuration options</b>
-</summary>
-The rollup plugin also takes an optional <code>interceptor</code>, which can be useful to handle things like rewriting <code>api.</code> prefixes made in requests, for example:
-
-```js
-mockRollupPlugin({
-  interceptor: `
-    const domain = window.location.hostname;
-    const apiDomain = "api." + domain;
-
-    const { fetch: originalFetch } = window;
-
-    async function fetch(request) {
-      request = new Request(request);
-
-      const resolvedURL = new URL(request.url, window.location);
-      if (resolvedURL.hostname === apiDomain) {
-        // rewrite hostname without api.
-        resolvedURL.hostname = domain;
-
-        const init = {};
-        for (const property in request) {
-          init[property] = request[property];
-        }
-
-        request = new Request(resolvedURL.href, init);
-        if (request.body != null) {
-          request = new Request(request, { body: await request.arrayBuffer() });
-        }
-      }
-      return originalFetch(request);
-    }
-
-    window.fetch = fetch;
-  `,
-});
-```
-
-This can be used to avoid CORS issues when deploying your Storybooks.
-
-In the Storybook 7+ (@web/storybook-builder) you can achieve the same by using native Storybook API [previewHead](https://storybook.js.org/docs/api/main-config/main-config-preview-head):
-
-```js
-// .storybook/main.js
-/** @type { import('@web/storybook-framework-web-components').StorybookConfig } */
-const config = {
-  framework: {
-    name: '@web/storybook-framework-web-components',
-  },
-  // ...
-  previewHead(head) {
-    return `
-      ${process.env.NODE_ENV === 'production' ? `<script>${interceptor}</script>` : ''}
-      ${head}
-    `;
-  },
-};
-export default config;
-```
-
-</details>
-<br/>
-
-And add the addon:
-`feature-a/.storybook/main.cjs`:
-
-```diff
-module.exports = {
-  stories: ['../stories/**/*.stories.{js,md,mdx}'],
-+  // for Storybook 7+ (@web/storybook-builder)
-+  addons: ['@web/mocks/storybook-addon'],
-
-+  // for Storybook 6 (@web/dev-server-storybook)
-+  addons: ['@web/mocks/storybook/addon.js'],
-  rollupConfig: async config => {
-    const { mockRollupPlugin } = await import('@web/mocks/plugins.js');
-    config.plugins.push(mockRollupPlugin());
-    return config;
-  },
-};
-```
-
-`feature-a/.storybook/preview.js`:
-
-```diff
-+// for Storybook 7+ (@web/storybook-builder)
-+// no need to do anything here
-
-+// for Storybook 6 (@web/dev-server-storybook)
-+ import { withMocks } from '@web/mocks/storybook/decorator.js';
-+ export const decorators = [withMocks];
-```
-
-`feature-a/stories/default.stories.js`:
-
-```js
-import { html } from 'lit';
-import { http } from '@web/mocks/http.js';
-import mocks from '../demo/mocks.js';
-
-export const Default = () => html`<feature-a></feature-a>`;
-Default.story = {
-  parameters: {
-    mocks: mocks.default,
-    // or
-    mocks: [
-      mocks.default,
-      otherMocks.error,
-      http.get('/api/bar', () => Response.json({ bar: 'bar' })),
-      http.post('/api/baz', () => new Response('', { status: 400 })),
-    ],
-    // or
-    mocks: [
-      http.get('/api/users/:id', ({ params }) => {
-        if (params.id === '123') {
-          return Response.json({ name: 'frank' });
-        }
-        return Response.json({ name: 'bob' });
-      }),
-    ],
-  },
-};
-```
 
 ## `@web/test-runner`
