--- packages/mocks/demo/wc/.storybook/main.js
@@ -1,9 +1,0 @@
-/** @type { import('../../../../index.d.ts').StorybookConfig } */
-const config = {
-  stories: ['../stories/**/*.stories.js'],
-  addons: ['@web/mocks/storybook-addon'],
-  framework: {
-    name: '@web/storybook-framework-web-components',
-  },
-};
-export default config;
