+++ packages/mocks/package.json
@@ -64,11 +64,6 @@
     "msw"
   ],
   "dependencies": {
-    "@storybook/manager-api": "^7.0.0 || ^8.0.0",
-    "@storybook/preview-api": "^7.0.0 || ^8.0.0",
-    "@web/storybook-prebuilt": "^0.1.37",
-    "@web/storybook-utils": "^1.1.1",
-    "lit": "^2.7.5 || ^3.0.0",
     "msw": "2.4.1"
   },
   "devDependencies": {
