+++ packages/mocks/plugins.js
@@ -1,2 +1,1 @@
-export { mockRollupPlugin } from './rollup-plugin.js';
 export { mockPlugin } from './wds-plugin.js';
