--- packages/mocks/storybook-addon/manager.js
@@ -1,1 +1,0 @@
-import '../storybook/addon/manager.js';
