--- packages/mocks/storybook-addon/preview.js
@@ -1,3 +1,0 @@
-import { withMocks } from '../storybook/addon/decorator.js';
-
-export const decorators = [withMocks];
