+++ packages/storybook-addon-mocks/CHANGELOG.md
@@ -1,0 +1,1 @@
+# @web/storybook-addons-mocks
