+++ packages/storybook-addon-mocks/package.json
@@ -1,0 +1,39 @@
+{
+  "name": "@web/storybook-addon-mocks",
+  "version": "0.0.0",
+  "publishConfig": {
+    "access": "public"
+  },
+  "description": "Storybook addon for `@web/mocks`",
+  "license": "MIT",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/modernweb-dev/web.git",
+    "directory": "packages/storybook-addon-mocks"
+  },
+  "author": "modern-web",
+  "homepage": "https://github.com/modernweb-dev/web/tree/master/packages/storybook-addon-mocks",
+  "type": "module",
+  "exports": {
+    "./manager": "./src/manager.js",
+    "./preset": "./src/preset.js",
+    "./preview": "./src/preview.js"
+  },
+  "files": [
+    "CHANGELOG.md",
+    "src/*.js",
+    "README.md"
+  ],
+  "keywords": [
+    "storybook",
+    "addon",
+    "mocks",
+    "msw"
+  ],
+  "dependencies": {
+    "@web/mocks": "^1.3.2",
+    "@web/storybook-utils": "^1.1.2",
+    "lit": "^3.3.2",
+    "storybook": "^9.1.19"
+  }
+}
