+++ packages/storybook-framework-web-components/tests/fixtures/all-in-one/.storybook/main.js
@@ -7,7 +7,7 @@ const config = {
     '@storybook/addon-docs',
     '@storybook/addon-links',
     '@storybook/addon-a11y',
-    '@web/mocks/storybook-addon',
+    '@web/storybook-addon-mocks',
   ],
   framework: {
     name: '@web/storybook-framework-web-components',
